backend/requirements.txt
fastapi[all]==0.104.1
uvicorn==0.24.0
langchain==0.0.350
openai==1.3.5
python-dotenv==1.0.0
supabase==2.3.0
pydantic==2.5.2
python-multipart==0.0.6
requests==2.31.0
chromadb==0.4.22
pypdf==3.17.1
tiktoken==0.5.2
langchain-community==0.0.13 